Move some of the packet digest processing to analyzer class

+++ trunk/AnalyzerCapture/src/PayloadLengthResult.cpp
@@ -66,7 +66,7 @@ ResultEnum CPayloadLengthResult::PrintResult()
 	return rs;
 }
 
-ResultEnum CPayloadLengthResult::PrintPacketLengthToFile(unsigned int* pArray)
+ResultEnum CPayloadLengthResult::PrintPacketLengthToFile(unsigned long long* pArray)
 {
 	if (NULL == pArray)
 	{
